Initial commit.
We are back on track!

This is the first commit of the new NPFC, which will substitute the old
„moreblocks“ stuff.

Type „make all“ in the root directory, and get all examples in the
examples(-sub-)directory running.

And yes, it is my plan to add html as output format also! But later!