Some minor fixes
**moreblocks.py:**
German Umlaute and \label doesn’t work together as expected. Now we use
an translation table as workaround. But we need to know if we use
pdflatex (which causes the problems) or LuaLaTeX/XeTeX (which will work
as expected) in future releases and handle this mistake properly.

Fixed the problem, that no new header behind the last moreblock-header
causes an error.

**splitbockold.py:**
splitbockold.py now has an extra \vspace* -1em at the beginning